build: Move include file generation to generate function
This has the advantage of letting the caller handle file I/O, so we only
need to care about the file contents itself. It also forces it to be
self-contained and not rely on whether we're using the OUT_PATH env
variable or an override through options (because yes, `include!` is
relative to the current file, so `./...` works in both cases).